feat: Added logic for play controllers (#20) Co-authored by @gegiske and @piapir

+++ core/src/main/java/com/fatpiggies/game/controller/IPlayController.java
@@ -3,5 +3,5 @@ package com.fatpiggies.game.controller;
 public interface IPlayController {
     void startGame(String lobbyId);
     void endGame(String lobbyId);
-    void movePig(int x, int y);
+    void movePig(double x, double y);
 }
